+++ CHANGELOG.md
@@ -22,6 +22,7 @@ To be compatible with Postbag 0.4 use `cfg::Version::Postbag0_4`.
   of values that failed deserialization either by their default or custom
   replacement values.
 - Optional header with Postbag version (enabled by default)
+- The `fixint` wrapper type is now public as `fixint::Fixint`
 - Minimum supported Rust version (MSRV) is 1.95
 
 ## 0.4.3